requirements.txt
langchain==0.0.299
gpt4all==1.0.8
chromadb==0.4.7
llama-cpp-python==0.2.6
urllib3==2.0.4
PyMuPDF==1.23.1
python-dotenv==1.0.0
unstructured==0.10.8
extract-msg==0.45.0
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.66.1
sentence_transformers==2.2.2
